fix: The function files were not working require bug fixed
After reordering controllers the error reqire is not defined in the module scope was thrown.

Fixed with adding log.js file content at the beginning of the controller files that have been thorwing error.

Do not merge with develop yet testing on apis will be done and a new commit will be made