feat: Serach bar is added to dashboard
Search bar will allow users to look for mathes and teams with the desired substring, number of elements shown to users is limeted to 5 on purpose it can be in creased in fact all the data can be shown. There are bugs which are related with issues #50 adn #52. Merge the branch with develop for this spreint but in the other sprint this needs fixing and possibly reverting.