docs: Function details are added as comments
Check the commetns and report any missing part after that run jsdoc <filename> to crate a API documentation for your functions. Follow this: 1- add your comment like what is added to matchController and in root folder run jsdoc <dir name> be sure to run all the files that include documentation comment and be sure to delete the already existing out folder hence the new one can replace it easily.