Chore: Match search is now connected to match page.
When clicked on match in search bar the user is sent to the match page. Before using the feture must add the check function
from video-scraping branch issue #67 it should filter the if statement as necessary. Also some style changes are made to the
search bar. Referee redirection is also changed too it will redirect to the proper page.