Add selecting feature and did some adjustements to the page
Before this commit, in both referees and matches we could have only seen the data but know we can interact with it also by choosing 1 from each. In the home segment of the page we see the selected referee and match, if there aren't any data it asks you to select one from the dedicated page segments. After selection there is a button to reset the values.